fix(library-solidity): sync bridge wrappers with updated host send/quote ABI
host #2880 dropped the options parameter from ConfidentialBridge.send/quote and narrowed lzComposeGas to uint64. Update the IConfidentialBridge mirror, the FHE.bridge/quoteBridge overloads (incl. the low-level bytes32 escape hatch), Impl, and the ConfidentialOAppSender helper accordingly, plus the example/harness/mocks/tests. The bridge now always builds its own LayerZero options and enforces a per-dstEid minimum compose gas.

+++ library-solidity/examples/bridge/OAppHarness.sol
@@ -45,7 +45,7 @@ contract OAppHarness is ConfidentialOAppSender, ConfidentialOAppReceiver {
         uint32 dstEid,
         bytes calldata payload,
         bytes32 handle,
-        uint128 lzComposeGas
+        uint64 lzComposeGas
     ) external payable returns (MessagingReceipt memory) {
         return _bridge(dstEid, payload, euint64.wrap(handle), lzComposeGas, msg.value);
     }
@@ -55,7 +55,7 @@ contract OAppHarness is ConfidentialOAppSender, ConfidentialOAppReceiver {
         uint32 dstEid,
         bytes calldata payload,
         bytes32 handle,
-        uint128 lzComposeGas
+        uint64 lzComposeGas
     ) external view returns (MessagingFee memory) {
         return _quoteBridge(dstEid, payload, euint64.wrap(handle), lzComposeGas);
     }
